requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
dateparser==0.7.0
Django==2.0.2
django-bower==5.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
krakenex==2.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-binance==0.6.6
python-dateutil==2.6.1
pytz==2018.3
regex==2018.2.21
requests==2.18.4
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
tzlocal==1.5.1
urllib3==1.22
zope.interface==4.4.3
